removing useEditorState from code editor, step 2
